requirements.txt
# Core
requests==2.31.0
python-dotenv==1.0.0

# Data processing
pandas==2.1.4
numpy==1.26.2

# Web scraping
beautifulsoup4==4.12.2
lxml==4.9.4
selenium==4.16.0

# API clients
anthropic==0.25.0  # Claude API (если будем использовать)

# Data validation
pydantic==2.5.3

# Logging
loguru==0.7.2

# Rate limiting
ratelimit==2.2.1

# Utilities
tqdm==4.66.1
python-slugify==8.0.1

# Testing (optional)
pytest==7.4.3
pytest-cov==4.1.0
